+++ lib/modules/group/index.ts
@@ -1,0 +1,4 @@
+export * from "./collections/groupsMappings";
+export * from "./GroupsModule";
+export * from "./roles/RoleGroupsAdmin";
+export * from "./roles/RoleGroupsReader";