+++ lib/modules/group/roles/RoleGroupsReader.ts
@@ -1,13 +1,13 @@
 import { KuzzleRole } from "../../shared/types/KuzzleRole";
 
 /**
- * This role allows to manage assets and their models.
+ * This role allows to manage groups and their models.
  *
  * It's a tenant role.
  *
  * @example
  *
-    "device-manager/assets": {
+    "device-manager/groups": {
       actions: {
         "*": true,
       },
@@ -22,14 +22,20 @@ import { KuzzleRole } from "../../shared/types/KuzzleRole";
       },
     },
  */
-export const RoleAssetsGroupsReader: KuzzleRole = {
-  name: "assetsGroup.reader",
+export const RoleGroupsReader: KuzzleRole = {
+  name: "group.reader",
   definition: {
     controllers: {
-      "device-manager/assetsGroup": {
+      "device-manager/groups": {
         actions: {
           get: true,
           search: true,
+        },
+      },
+      "device-manager/models": {
+        actions: {
+          listGroups: true,
+          getGroup: true,
         },
       },
     },